Add very basic add/remove list functionality

+++ app/Http/Controllers/ListItemsController.php
@@ -3,10 +3,18 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use App\ListItem;
 
 //These will never return anything to the user, rather, these will likely just be called from ajax requests.
 class ListItemsController extends Controller
 {
+
+    //Need auth to access anything from this controlelr
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -14,7 +22,8 @@ class ListItemsController extends Controller
      */
     public function index()
     {
-        //
+        //this might actually be what homecontroller should call.
+        //Should this return an view, or just all the data?
     }
 
     /**
@@ -25,7 +34,7 @@ class ListItemsController extends Controller
      */
     public function create()
     {
-        //
+        // Should not be accessable in usual operation.
     }
 
     /**
@@ -40,33 +49,52 @@ class ListItemsController extends Controller
             'priority' => 'nullable',
             'item_desc' => 'required'
         ]);
-        //
+
+        $listItem = new ListItem();
+        $listItem->priority = $request->input('priority');
+        $listItem->item_desc = $request->input('item_desc');
+        $listItem->user_id = auth()->user()->id;
+        $listItem->save();
+
+        return redirect('/home')->with('success', 'Item Added.');
+        // return json_encode(array('status' => 'success', 'message' => 'Item Added.'));
+
     }
+
+    // public function retrieveAll(){
+
+    // }
 
     /**
      * Display the specified resource.
+     * 
+     * Returns a response to AJAX.
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
     {
-        //
+        // Should not be accessable
     }
 
     /**
      * Show the form for editing the specified resource.
+     * 
+     * Returns a response to AJAX
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
     {
-        //
+
     } 
 
     /**
      * Update the specified resource in storage.
+     * 
+     * Returns a response to AJAX.
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
@@ -74,11 +102,31 @@ class ListItemsController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        $this->validate($request,[
+            'priority' => 'nullable',
+            'item_desc' => 'required'
+        ]);
+
+        $listItem = ListItem::find($id);
+
+        if(auth()->user->id !== $listItem->user_id){
+            return json_encode(array('status' => 'error', 'message' => 'Unauthorized Request'));
+        }
+
+        $listItem->priority = $request->input('priority');
+        $listItem->item_desc = $request->input('item_desc');
+        $listItem->save();
+
+        return redirect('/home')->with('success', 'Item Updated.');
+
+        // return json_encode(array('status' => 'success', 'message' => 'Updated Item.'));
+
     }
 
     /**
      * Remove the specified resource from storage.
+     * 
+     * Returns a response to AJAX
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
@@ -87,7 +135,14 @@ class ListItemsController extends Controller
     {
         $listItem = ListItem::find($id);
 
+        if(auth()->user()->id !== $listItem->user_id){
+            return json_encode(array('status' => 'error', 'message' => 'Unauthorized Request'));
+        }
 
         $listItem->delete();
+
+        return redirect('/home')->with('success', 'Item Removed');
+
+        // return json_encode(array('status' => 'success', 'message' => 'Item Deleted.'));
     }
 }
